Restructuring code in packages
